Add aliases.
Note: There are different definitions of poissons ratio!
(VDI 2014 Blatt 3 page 14)
In case of the Poisson’s ratios there are different ways
for the indexing in international practice. In the
guideline VDI 2014 the required two indices are uti-
lized as follows: The 1 st index indicates the direction
of the transverse contraction. The 2 nd index denotes
the stress, which causes the contraction. As a conse-
quence the Poisson’s ratios ν ⊥|| is the larger and ν ||⊥
the smaller one. (In the English literature the two indices related to 
the contraction and acting stress are
used in the reverse sequence.)