Clarify 'principal axis':
Using 'principal_axis' indicates, that whether E1 is Et or El depends on 
selection of 'principal axis'. However, this is not the case. E1 is 
always used as El and after having created the stiffness, the stiffness 
is rotated to fit the direction specified by 'rotate_principal_axis_to'.

This wording is not optimal.
Maybe one should use only El, Et, Glt, Gtt, nult, ...